+++ packages/cldr-core/src/internals/internals.ts
@@ -60,8 +60,8 @@ export interface DateFieldInternals {
 export interface GeneralInternals {
   characterOrder(bundle: Bundle): string;
   lineOrder(bundle: Bundle): string;
-  contextTransform(value: string, context: ContextType,
-    info: ContextTransformInfo, field?: ContextTransformFieldType): string;
+  contextTransform(value: string, info: ContextTransformInfo,
+    context?: ContextType, field?: ContextTransformFieldType): string;
   formatList(bundle: Bundle, items: string[], type: ListPatternType): string;
   formatListToParts(bundle: Bundle, items: string[], type: ListPatternType): Part[];
   formatListToPartsImpl(bundle: Bundle, items: Part[][], type: ListPatternType): Part[];