+++ packages/cldr-schema/src/code/datefields.ts
@@ -1,5 +1,5 @@
 import { scope, vector1, vector2, Instruction, Scope } from '../types';
-import { DateFieldIndex, RelativeTimeFieldIndex } from '../schema';
+import { DateFieldIndex, DateFieldWidthIndex, RelativeTimeFieldIndex } from '../schema';
 
 const prevNext: Instruction[] = ['previous2', 'previous', 'current', 'next', 'next2']
   .map(k => vector1(k, 'relative-time-field'));
@@ -7,9 +7,7 @@ const prevNext: Instruction[] = ['previous2', 'previous', 'current', 'next', 'ne
 const futurePast: Instruction[] = ['future', 'past']
   .map(k => vector2(k, 'plural-key', 'relative-time-field'));
 
-const displayName = vector1('displayName', 'date-field');
-
-const relativeTimeBody = prevNext.concat(futurePast).concat([displayName]);
+const relativeTimeBody = prevNext.concat(futurePast);
 
 const relativeTimes = (width: string) => scope(width, width, relativeTimeBody);
 
@@ -19,9 +17,12 @@ export const DATEFIELDS: Scope = scope('DateFields', 'DateFields', [
     relativeTimes('short'),
     relativeTimes('narrow'),
   ]),
+
+  vector2('displayName', 'date-field', 'date-field-width')
 ]);
 
 export const DATEFIELDS_INDICES = {
   'date-field': DateFieldIndex,
+  'date-field-width': DateFieldWidthIndex,
   'relative-time-field': RelativeTimeFieldIndex
 };